Carve `TestDatasets` out of the monolith (wala/ML#635).
The 105 dataset, collection-probe, and generator-iteration tests move
verbatim; the monolith keeps 592, preserving the 1,040 total.
